refactor: Rename classes to match standards
