+++ Scripts/inject/InjectUtils.js
